requirements.txt

tensorflow==2.13.1
numpy>=1.23,<2.0
pandas>=1.5,<2.1
pillow>=9.5
streamlit>=1.32,<1.36
opencv-python-headless>=4.8
matplotlib>=3.7
scikit-learn>=1.3
